--- oop-application/src/main/kotlin/com/jaeyong/oop/application/post/service/CreatePostService.kt
@@ -1,26 +1,0 @@
-package com.jaeyong.oop.application.post.service
-
-import com.jaeyong.oop.application.post.command.CreatePostCommand
-import com.jaeyong.oop.application.post.result.CreatePostResult
-import com.jaeyong.oop.application.post.usecase.CreatePostUseCase
-import com.jaeyong.oop.domain.post.Post
-import com.jaeyong.oop.domain.post.PostContent
-import com.jaeyong.oop.domain.post.PostTitle
-import com.jaeyong.oop.domain.post.port.PostPort
-import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
-
-/** VO 생성으로 도메인 검증 후 저장 */
-@Service
-class CreatePostService(private val postPort: PostPort) : CreatePostUseCase {
-
-    @Transactional
-    override fun create(command: CreatePostCommand): CreatePostResult {
-        val post = Post.create(
-            title = PostTitle.of(command.title),
-            content = PostContent.of(command.content),
-            authorUsername = command.authorUsername
-        )
-        return CreatePostResult.of(postPort.save(post))
-    }
-}
