--- oop-application/src/main/kotlin/com/jaeyong/oop/application/post/usecase/DeletePostUseCase.kt
@@ -1,7 +1,0 @@
-package com.jaeyong.oop.application.post.usecase
-
-import com.jaeyong.oop.application.post.command.DeletePostCommand
-
-interface DeletePostUseCase {
-    fun delete(command: DeletePostCommand)
-}
