--- oop-application/src/main/kotlin/com/jaeyong/oop/application/post/usecase/GetPostListUseCase.kt
@@ -1,8 +1,0 @@
-package com.jaeyong.oop.application.post.usecase
-
-import com.jaeyong.oop.application.post.command.GetPostListCommand
-import com.jaeyong.oop.application.post.result.GetPostListResult
-
-interface GetPostListUseCase {
-    fun getList(command: GetPostListCommand): GetPostListResult
-}
