--- oop-application/src/main/kotlin/com/jaeyong/oop/application/post/usecase/GetPostUseCase.kt
@@ -1,8 +1,0 @@
-package com.jaeyong.oop.application.post.usecase
-
-import com.jaeyong.oop.application.post.command.GetPostCommand
-import com.jaeyong.oop.application.post.result.GetPostResult
-
-interface GetPostUseCase {
-    fun get(command: GetPostCommand): GetPostResult
-}
