+++ oop-presentation/src/main/kotlin/com/jaeyong/oop/presentation/post/api/PostController.kt
@@ -5,11 +5,7 @@ import com.jaeyong.oop.application.post.command.DeletePostCommand
 import com.jaeyong.oop.application.post.command.GetPostCommand
 import com.jaeyong.oop.application.post.command.GetPostListCommand
 import com.jaeyong.oop.application.post.command.UpdatePostCommand
-import com.jaeyong.oop.application.post.usecase.CreatePostUseCase
-import com.jaeyong.oop.application.post.usecase.DeletePostUseCase
-import com.jaeyong.oop.application.post.usecase.GetPostListUseCase
-import com.jaeyong.oop.application.post.usecase.GetPostUseCase
-import com.jaeyong.oop.application.post.usecase.UpdatePostUseCase
+import com.jaeyong.oop.application.post.usecase.PostUseCase
 import com.jaeyong.oop.presentation.auth.CurrentUser
 import com.jaeyong.oop.presentation.post.request.CreatePostRequest
 import com.jaeyong.oop.presentation.post.request.UpdatePostRequest
@@ -25,11 +21,7 @@ import org.springframework.web.bind.annotation.*
 @RestController
 @RequestMapping("/posts")
 class PostController(
-    private val createPostUseCase: CreatePostUseCase,
-    private val getPostUseCase: GetPostUseCase,
-    private val getPostListUseCase: GetPostListUseCase,
-    private val updatePostUseCase: UpdatePostUseCase,
-    private val deletePostUseCase: DeletePostUseCase
+    private val postUseCase: PostUseCase
 ) {
 
     @PostMapping
@@ -37,13 +29,13 @@ class PostController(
         @Valid @RequestBody request: CreatePostRequest,
         @CurrentUser username: String
     ): ResponseEntity<ApiResponse<PostResponse>> {
-        val result = createPostUseCase.create(CreatePostCommand.of(request.title, request.content, username))
+        val result = postUseCase.create(CreatePostCommand.of(request.title, request.content, username))
         return ApiResponse.success(PostResponse.of(result), HttpStatus.CREATED)
     }
 
     @GetMapping("/{id}")
     fun get(@PathVariable id: Long): ResponseEntity<ApiResponse<PostResponse>> {
-        val result = getPostUseCase.get(GetPostCommand.of(id))
+        val result = postUseCase.get(GetPostCommand.of(id))
         return ApiResponse.success(PostResponse.of(result), HttpStatus.OK)
     }
 
@@ -52,7 +44,7 @@ class PostController(
         @RequestParam(defaultValue = "0") page: Int,
         @RequestParam(defaultValue = "10") size: Int
     ): ResponseEntity<ApiResponse<PostListResponse>> {
-        val result = getPostListUseCase.getList(GetPostListCommand.of(page, size))
+        val result = postUseCase.getList(GetPostListCommand.of(page, size))
         return ApiResponse.success(PostListResponse.of(result), HttpStatus.OK)
     }
 
@@ -62,7 +54,7 @@ class PostController(
         @Valid @RequestBody request: UpdatePostRequest,
         @CurrentUser username: String
     ): ResponseEntity<ApiResponse<PostResponse>> {
-        val result = updatePostUseCase.update(UpdatePostCommand.of(id, request.title, request.content, username))
+        val result = postUseCase.update(UpdatePostCommand.of(id, request.title, request.content, username))
         return ApiResponse.success(PostResponse.of(result), HttpStatus.OK)
     }
 
@@ -71,7 +63,7 @@ class PostController(
         @PathVariable id: Long,
         @CurrentUser username: String
     ): ResponseEntity<ApiResponse<Nothing>> {
-        deletePostUseCase.delete(DeletePostCommand.of(id, username))
+        postUseCase.delete(DeletePostCommand.of(id, username))
         return ApiResponse.success(HttpStatus.NO_CONTENT)
     }
 }
